# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PIX Payment Integration API",
    "description": "Complete PIX (Brazilian instant payment) integration for Love to Fly Portal. Supports hangar bookings and membership upgrades with webhook and reconciliation mechanisms.",
    "version": "1.0.0",
    "contact": {
      "name": "Love to Fly Team",
      "email": "support@lovetofly.com"
    }
  },
  "servers": [
    {
      "url": "https://lovetofly.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from user authentication"
      },
      "CronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Secret for cron job reconciliation"
      },
      "WebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "asaas-access-token",
        "description": "Webhook token configured in Asaas (asaas-access-token header)"
      }
    },
    "schemas": {
...
      "BRCode": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "Base64 encoded QR code image"
          },
          "brCode": {
            "type": "string",
            "description": "Raw PIX BRCode string for copy-paste"
          },
          "txId": {
            "type": "string",
            "description": "Transaction ID (matches order_id for reconciliation)"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
  "paths": {
    "/api/payments/pix/create": {
      "post": {
        "tags": ["PIX Payment Creation"],
        "summary": "Create a new PIX payment",
        "description": "Initiates a PIX payment order and generates a BRCode for user scanning. Returns QR code and raw string for copy-paste.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderType", "orderId", "amountCents"],
                "properties": {
                  "orderType": {
                    "type": "string",
                    "enum": ["hangar_booking", "membership"],
                    "description": "Type of order"
                  },
                  "orderId": {
                    "type": "string",
                    "example": "booking-123",
                    "description": "Unique order reference (e.g., 'booking-123', 'membership-pro')"
                  },
                  "amountCents": {
                    "type": "integer",
                    "example": 50000,
                    "description": "Amount in cents (50000 = R$ 500.00)"
                  },
                  "description": {
                    "type": "string",
                    "example": "Hangar booking for 3 months",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "order_id": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "object",
                      "properties": {
                        "cents": {
                          "type": "integer"
                        },
                        "formatted": {
                          "type": "string"
                        }
                      }
                    },
                    "brCode": {
                      "$ref": "#/components/schemas/BRCode"
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/pix-reconcile": {
      "post": {
        "tags": ["Admin Tools"],
        "summary": "Admin reconciliation tool",
        "description": "Manual reconciliation endpoint for administrators. Marks expired payments and confirms pending bookings.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "confirmPendingBookings": {
                    "type": "boolean",
                    "default": false,
                    "description": "Also confirm pending bookings with completed payments"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin reconciliation completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "processed": {
                      "type": "integer"
                    },
                    "skipped": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "pending": {
                          "type": "integer"
                        },
                        "completed": {
                          "type": "integer"
                        },
                        "expired": {
                          "type": "integer"
                        },
                        "failed": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - admin role required"
          }
        }
      }
    },
...
  },
...
}